WIP: Add root object to gizmo as alternative anchor for positioning and orientation
The root object is currently disabled / commented out as there is a problem with initializing its value. Setting the root object outside of the constructor seems to break the draw function calls (draw will only be called after releasing the gizmo handle, not while dragging). Initialization in constructor is not possible as the hierarchy has to be traversed.