Fix bad return type in transforming interface
The const mat& return type was a leftover from when transforming still always had a matrix as a member variable. For the current version a const reference does not make sense. A shared_ptr might be possible instead of by value but it's much safer if the matrix cannot change (i.e. is an actual copy).