Add ref-relations and source locations to doxy-doc
Each documented function now contains an auto-generated link to all
functions it uses and functions where it is used. Additionally the
definition location of the function is provided.

This change might result, on some instances, in a quite ugly and verbose
list but overall highly improves the symbol navigation experience in
the documentation. Oftentimes cgv functions can only be understood from
their usage patterns. And these are now much easier to discover.