Disambiguate dynamic_mesh declarations as types
Although the code compiles on msvc and clang, it seems that gcc has some
issues. The immediate base class simple_mesh<T> is templated and
therefore seems to make the simple_mesh_base::vec2i declaration a
dependently looked-up name. Therefore the typename keyword needs to be
present to ensure the compiler that we have name a type with the
succeeding declaration and not something else ...

See the following links for more info:

[1] https://stackoverflow.com/questions/1643035/propagating-typedef-from-based-to-derived-class-for-template
[2] https://isocpp.org/wiki/faq/templates#nondependent-name-lookup-types