Add shader file name cache.
This improves shader loading times during development for DLL builds. Every shader file that is not found as a resource is searched recursively
in the shader paths, which performs many slow file system queries. Using the cache, all paths are now scanned once and all file names with extension
matching .gl* are recorded to the cache. This allows for fast lookup of the full shader path for a given file name, avoiding the recursive search.