Rollback to cf1d523cb4c814bba17bd33379cc263585238bdb and replace changes with proper unified cleanup behavior via new gui_driver::destroy_all_windows API.
Problem:
On Windows, the release of some context-held OpenGL resources causes a crash on shutdown. For to-be-investigated reasons, this does not happen when closing the main Window via ESC or the close button, likely because FLTK cleans up the main Window explicitly (see below) in this case.

Cause:
Unknown. However, it seems to be related to FLTK performing cleanup of window resources lazily when the module gets unloaded by the OS. It seems the OpenGL context becomes invalid by that time, although the exact reason for this is unclear. Furthermore, this appears to be an issue only on Windows (perhaps FLTK performs eager cleanup on X11, to-be-investigated).

Fix:
Perform explicit cleanup before leaving the main function. Leverage the new destroy_all_windows API in the gui_driver for this.